+++ gradle/projects.main.properties
@@ -30,8 +30,8 @@ polaris-quarkus-service=quarkus/service
 polaris-quarkus-server=quarkus/server
 polaris-quarkus-spark-tests=quarkus/spark-tests
 polaris-quarkus-admin=quarkus/admin
-polaris-quarkus-commons=quarkus/commons
-polaris-quarkus-test-commons=quarkus/test-commons
+polaris-quarkus-common=quarkus/common
+polaris-quarkus-test-common=quarkus/test-common
 polaris-quarkus-run-script=quarkus/run-script
 polaris-eclipselink=extension/persistence/eclipselink
 polaris-jpa-model=extension/persistence/jpa-model
