+++ quarkus/admin/build.gradle.kts
@@ -53,8 +53,8 @@ dependencies {
 
   implementation("org.jboss.slf4j:slf4j-jboss-logmanager")
 
-  implementation(project(":polaris-quarkus-commons"))
-  testImplementation(project(":polaris-quarkus-test-commons"))
+  implementation(project(":polaris-quarkus-common"))
+  testImplementation(project(":polaris-quarkus-test-common"))
   testFixturesApi(project(":polaris-core"))
 
   testFixturesApi(enforcedPlatform(libs.quarkus.bom))
