+++ quarkus/service/src/intTest/java/org/apache/polaris/service/quarkus/it/relational/jdbc/JdbcQuarkusApplicationIT.java
@@ -21,7 +21,7 @@ package org.apache.polaris.service.quarkus.it.relational.jdbc;
 import io.quarkus.test.junit.QuarkusIntegrationTest;
 import io.quarkus.test.junit.TestProfile;
 import org.apache.polaris.service.it.test.PolarisApplicationIntegrationTest;
-import org.apache.polaris.test.commons.RelationalJdbcProfile;
+import org.apache.polaris.test.common.RelationalJdbcProfile;
 
 @TestProfile(RelationalJdbcProfile.class)
 @QuarkusIntegrationTest
