+++ quarkus/test-common/build.gradle.kts
@@ -16,22 +16,27 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.test.commons;
 
-import io.quarkus.test.junit.QuarkusTestProfile;
-import java.util.List;
-import java.util.Map;
+plugins {
+  alias(libs.plugins.jandex)
+  id("java-test-fixtures")
+}
 
-public class RelationalJdbcProfile implements QuarkusTestProfile {
-  @Override
-  public Map<String, String> getConfigOverrides() {
-    return Map.of("polaris.persistence.auto-bootstrap-types", "relational-jdbc");
-  }
+configurations.all {
+  exclude(group = "org.antlr", module = "antlr4-runtime")
+  exclude(group = "org.scala-lang", module = "scala-library")
+  exclude(group = "org.scala-lang", module = "scala-reflect")
+}
 
-  @Override
-  public List<TestResourceEntry> testResources() {
-    return List.of(
-        new QuarkusTestProfile.TestResourceEntry(
-            PostgresRelationalJdbcLifeCycleManagement.class, Map.of()));
-  }
+java {
+  sourceCompatibility = JavaVersion.VERSION_21
+  targetCompatibility = JavaVersion.VERSION_21
+}
+
+dependencies {
+  implementation(enforcedPlatform(libs.quarkus.bom))
+  implementation("io.quarkus:quarkus-junit5")
+  implementation(platform(libs.testcontainers.bom))
+  implementation("org.testcontainers:testcontainers")
+  implementation("org.testcontainers:postgresql")
 }
