+++ quarkus/test-common/src/main/java/org/apache/polaris/test/common/PostgresRelationalJdbcLifeCycleManagement.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.test.commons;
+package org.apache.polaris.test.common;
 
 import io.quarkus.test.common.DevServicesContext;
 import io.quarkus.test.common.QuarkusTestResourceLifecycleManager;
@@ -64,7 +64,7 @@ public class PostgresRelationalJdbcLifeCycleManagement
     if (databases.isEmpty()) {
       Map<String, String> props = generateDataSourceProps(List.of("realm1"));
       // make realm1_ds as the default ds
-      props.put("polaris.relation.jdbc.datasource.default-datasource", "realm1_ds");
+      props.put("polaris.relational.jdbc.datasource.default-datasource", "realm1_ds");
       return props;
     } else {
       Map<String, String> allProps = new HashMap<>();
@@ -97,7 +97,7 @@ public class PostgresRelationalJdbcLifeCycleManagement
     Map<String, String> props = new HashMap<>();
     // polaris.relation.jdbc.datasource.realm=realm_ds
     for (String database : realms) {
-      props.put(String.format("polaris.relation.jdbc.datasource.%s", database), database + "_ds");
+      props.put(String.format("polaris.relational.jdbc.datasource.%s", database), database + "_ds");
     }
     return props;
   }
